+++ internal/processor/filterexpr/matcher.go
@@ -50,8 +50,8 @@ func (m *Matcher) MatchMetric(metric pdata.Metric) (bool, error) {
 		return m.matchDoubleGauge(metricName, metric.DoubleGauge())
 	case pdata.MetricDataTypeIntSum:
 		return m.matchIntSum(metricName, metric.IntSum())
-	case pdata.MetricDataTypeDoubleSum:
-		return m.matchDoubleSum(metricName, metric.DoubleSum())
+	case pdata.MetricDataTypeSum:
+		return m.matchSum(metricName, metric.Sum())
 	case pdata.MetricDataTypeIntHistogram:
 		return m.matchIntHistogram(metricName, metric.IntHistogram())
 	case pdata.MetricDataTypeHistogram:
@@ -89,7 +89,7 @@ func (m *Matcher) matchDoubleGauge(metricName string, gauge pdata.DoubleGauge) (
 	return false, nil
 }
 
-func (m *Matcher) matchDoubleSum(metricName string, sum pdata.DoubleSum) (bool, error) {
+func (m *Matcher) matchSum(metricName string, sum pdata.Sum) (bool, error) {
 	pts := sum.DataPoints()
 	for i := 0; i < pts.Len(); i++ {
 		matched, err := m.matchEnv(metricName, pts.At(i).LabelsMap())