+++ translator/internaldata/metrics_to_oc_test.go
@@ -236,9 +236,9 @@ func TestMetricsType(t *testing.T) {
 			name: "double-non-monotonic-delta-sum",
 			internal: func() pdata.Metric {
 				m := pdata.NewMetric()
-				m.SetDataType(pdata.MetricDataTypeDoubleSum)
-				m.DoubleSum().SetIsMonotonic(false)
-				m.DoubleSum().SetAggregationTemporality(pdata.AggregationTemporalityDelta)
+				m.SetDataType(pdata.MetricDataTypeSum)
+				m.Sum().SetIsMonotonic(false)
+				m.Sum().SetAggregationTemporality(pdata.AggregationTemporalityDelta)
 				return m
 			},
 			descType: ocmetrics.MetricDescriptor_GAUGE_DOUBLE,
@@ -247,9 +247,9 @@ func TestMetricsType(t *testing.T) {
 			name: "double-non-monotonic-cumulative-sum",
 			internal: func() pdata.Metric {
 				m := pdata.NewMetric()
-				m.SetDataType(pdata.MetricDataTypeDoubleSum)
-				m.DoubleSum().SetIsMonotonic(false)
-				m.DoubleSum().SetAggregationTemporality(pdata.AggregationTemporalityCumulative)
+				m.SetDataType(pdata.MetricDataTypeSum)
+				m.Sum().SetIsMonotonic(false)
+				m.Sum().SetAggregationTemporality(pdata.AggregationTemporalityCumulative)
 				return m
 			},
 			descType: ocmetrics.MetricDescriptor_GAUGE_DOUBLE,
@@ -258,9 +258,9 @@ func TestMetricsType(t *testing.T) {
 			name: "double-monotonic-delta-sum",
 			internal: func() pdata.Metric {
 				m := pdata.NewMetric()
-				m.SetDataType(pdata.MetricDataTypeDoubleSum)
-				m.DoubleSum().SetIsMonotonic(true)
-				m.DoubleSum().SetAggregationTemporality(pdata.AggregationTemporalityDelta)
+				m.SetDataType(pdata.MetricDataTypeSum)
+				m.Sum().SetIsMonotonic(true)
+				m.Sum().SetAggregationTemporality(pdata.AggregationTemporalityDelta)
 				return m
 			},
 			descType: ocmetrics.MetricDescriptor_GAUGE_DOUBLE,
@@ -269,9 +269,9 @@ func TestMetricsType(t *testing.T) {
 			name: "double-monotonic-cumulative-sum",
 			internal: func() pdata.Metric {
 				m := pdata.NewMetric()
-				m.SetDataType(pdata.MetricDataTypeDoubleSum)
-				m.DoubleSum().SetIsMonotonic(true)
-				m.DoubleSum().SetAggregationTemporality(pdata.AggregationTemporalityCumulative)
+				m.SetDataType(pdata.MetricDataTypeSum)
+				m.Sum().SetIsMonotonic(true)
+				m.Sum().SetAggregationTemporality(pdata.AggregationTemporalityCumulative)
 				return m
 			},
 			descType: ocmetrics.MetricDescriptor_CUMULATIVE_DOUBLE,